refactor: move ledgerdb
